requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-cors-headers==3.4.0
djangorestframework==3.12.4
mock==4.0.3
pip==21.0.1
psycopg2==2.8.6
pytz==2021.1
setuptools==52.0.0
sqlparse==0.4.1
wheel==0.36.2
wincertstore==0.2
